face-emotion-recognition/requirements.txt
torch
torchvision
opencv-python
numpy
Pillow
gradio>=4.0